pyproject.toml
[tool.poetry]
name = "kanon-api"
version = "0.0.1"
description = "Kanon features exposed through an API"
authors = ["Léni Gauffier <leni.gauffier@obspm.fr>"]
readme = "README.md"
repository = "https://github.com/legau/kanon-api"
documentation = "https://kanon-api.readthedocs.io"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/legau/kanon-api/issues"
"Changelog" = "https://kanon-api.readthedocs.io/changelog.html"

[tool.poetry.dependencies]
python = "^3.9"
kanonpy = "^0.1.1"
fastapi = "^0.63.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.8.0"
pytest = "^6.2.3"
pytest-cov = "^2.10.1"
mypy = "^0.812"
pre-commit = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
known_first_party = ["app", "tests"]
